+++ src/level3ai.h
@@ -19,7 +19,7 @@ class TileNull;
 class Level3AI : public AI {
     int valueOfOccupant(const Tile&);
   protected:
-    std::vector<std::pair<int, std::pair<size_t, size_t>>> assignPriorities(
+    std::vector<std::pair<float, std::pair<size_t, size_t>>> assignPriorities(
       const std::vector<std::pair<size_t, size_t>> &, const GameState&) override;
     std::pair<size_t, size_t> findMove(const GameState&) override;
   public:
